Move embedding service to elasticdl/python/master
