+++ elasticdl/python/tests/optimizer_wrapper_test.py
@@ -5,8 +5,8 @@ import numpy as np
 import tensorflow as tf
 from tensorflow.keras.optimizers import SGD, Adam
 
-from elasticdl.python.common.embedding_service import EmbeddingService
 from elasticdl.python.elasticdl.layers.embedding import Embedding
+from elasticdl.python.master.embedding_service import EmbeddingService
 from elasticdl.python.master.optimizer_wrapper import (
     OptimizerWrapper,
     _parse_lookup_values,
